requirements.txt
# pip install -r requirements.txt
scipy==1.4.1
tqdm==4.32.1
numpy==1.19.2
pandas==1.2.5
scikit-learn==0.23.2
goatools==1.0.15
venn==0.1.3
termcolor==1.1.0
seaborn==0.11.1
matplotlib==3.3.4
matplotlib_venn==0.11.6
biopython==1.76
ipysheet==0.4.4
openpyxl==3.0.7
tabulate==0.8.9
gseapy==0.10.4
